backend/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
anthropic
python-dotenv
lxml
